Adjust format strings where off_t may be long long
The easiest thing to do to work on platforms where off_t may be
different widths is to expand the value being printed to be a long long,
and then use the long long format specifier.

+++ source/explore-index.cpp
@@ -35,8 +35,8 @@ static void walk(IndexReader const& reader, off_t node, int64_t count,
 
   for (size_t i = 0; i < children.size(); ++i) {
     sofar->push_back(children[i].ch);
-    printf("%s (%" PRId64 ") @%zd\n", sofar->c_str(),
-        children[i].count, children[i].next);
+    printf("%s (%" PRId64 ") @%lld\n", sofar->c_str(),
+        children[i].count, static_cast<long long>(children[i].next));
     walk(reader, children[i].next, children[i].count, path, depth - 1, sofar);
     sofar->resize(sofar->size() - 1);
   }
@@ -56,7 +56,8 @@ int main(int argc, char *argv[]) {
 
   IndexReader reader(input);
 
-  printf("Root (%" PRId64 ") @%zd\n", reader.count(), reader.root());
+  printf("Root (%" PRId64 ") @%lld\n", reader.count(),
+      static_cast<long long>(reader.root()));
 
   int depth = strlen(argv[2]);
   if (argc > 3) {
